requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
DateTime==5.5
idna==3.10
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
zope.interface==7.0.3
